requirements.txt
streamlit
python-dotenv
langchain>=0.1.17
langchain-core>=0.1.45
langchain-community>=0.1.17
langchain-google-genai>=1.0.3
google-generativeai>=0.4.1
faiss-cpu
pymupdf
